Split classes into separate files to cleanliness
